requirements.txt
torch
soundfile
scipy
transformers
ffmpeg-python
djangorestframework
pytube